Move SampleWindow to size_handle module; rename and share
