+++ kas-wgpu/src/theme/mod.rs
@@ -9,7 +9,7 @@ mod flat_theme;
 mod shaded_theme;
 mod size_handle;
 
-pub(crate) use size_handle::{Dimensions, DimensionsParams, SizeHandle};
+pub(crate) use size_handle::{Dimensions, DimensionsParams, DimensionsWindow};
 
 pub use flat_theme::FlatTheme;
 pub use shaded_theme::ShadedTheme;
